Updated file

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-		double temperatureFahrenheit = 9 * temperatureCelsius / 5 + 32;
+		double temperatureFahrenheit = 9 * (double)temperatureCelsius / 5 + 32;
 		System.out.println(temperatureFahrenheit);
     }
 }
